Remove useless isDefault for RequiredSPI impl (#16220)
