+++ shardingsphere-infra/shardingsphere-infra-datetime/shardingsphere-infra-datetime-type/shardingsphere-database-datetime/src/main/java/org/apache/shardingsphere/datetime/database/DatabaseDatetimeServiceDelegate.java
@@ -33,9 +33,4 @@ public final class DatabaseDatetimeServiceDelegate implements DatetimeService {
     public Date getDatetime() {
         return INSTANCE.getDatetime();
     }
-    
-    @Override
-    public boolean isDefault() {
-        return false;
-    }
 }
