+++ shardingsphere-infra/shardingsphere-infra-datetime/shardingsphere-infra-datetime-type/shardingsphere-database-datetime/src/main/java/org/apache/shardingsphere/datetime/database/TimeServiceConfiguration.java
@@ -31,10 +31,6 @@ import java.util.Properties;
 
 /**
  * Time service configuration.
- * 
- * <p>
- * Need to create a time-service.properties under the classpath.
- * </p>
  */
 @Getter
 public final class TimeServiceConfiguration {
