+++ shardingsphere-infra/shardingsphere-infra-datetime/shardingsphere-infra-datetime-type/shardingsphere-database-datetime/src/main/java/org/apache/shardingsphere/datetime/database/impl/DatabaseDatetimeService.java
@@ -49,9 +49,4 @@ public final class DatabaseDatetimeService implements DatetimeService {
         }
         return new Date();
     }
-    
-    @Override
-    public boolean isDefault() {
-        return false;
-    }
 }
