+++ shardingsphere-infra/shardingsphere-infra-datetime/shardingsphere-infra-datetime-type/shardingsphere-database-datetime/src/main/resources/META-INF/services/org.apache.shardingsphere.datetime.database.spi.DatabaseSQLEntry
@@ -15,7 +15,7 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.datetime.database.spi.MySQLDatabaseSQLEntry
-org.apache.shardingsphere.datetime.database.spi.PostgreSQLDatabaseSQLEntry
-org.apache.shardingsphere.datetime.database.spi.OracleDatabaseSQLEntry
-org.apache.shardingsphere.datetime.database.spi.SQLServerDatabaseSQLEntry
+org.apache.shardingsphere.datetime.database.spi.dialect.MySQLDatabaseSQLEntry
+org.apache.shardingsphere.datetime.database.spi.dialect.PostgreSQLDatabaseSQLEntry
+org.apache.shardingsphere.datetime.database.spi.dialect.OracleDatabaseSQLEntry
+org.apache.shardingsphere.datetime.database.spi.dialect.SQLServerDatabaseSQLEntry
