+++ shardingsphere-spi/src/test/java/org/apache/shardingsphere/spi/fixture/required/RequiredSPIFixtureDefaultFalseImpl.java
@@ -18,10 +18,4 @@
 package org.apache.shardingsphere.spi.fixture.required;
 
 public final class RequiredSPIFixtureDefaultFalseImpl implements RequiredSPIFixture {
-    
-    @Override
-    public boolean isDefault() {
-        return false;
-    }
 }
-
